refactor:AnalysisRoomのコンポーネント名修正
