+++ src/components/AnalisysRoom/AnalysisRoom.tsx
