+++ src/components/AnalisysRoom/Scatter2D.tsx
