+++ src/components/AnalisysRoom/TendScore.ts
