+++ src/components/AnalisysRoom/analysisUlits.spec.ts
