+++ src/components/AnalisysRoom/analysisUtils.ts
